# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Fil Application",
    "version": "1.0",
    "contact": {
      "name": "Hoang Ngoc Thach",
      "email": "hnt1693@gmail.com"
    },
    "description": "App cá độ"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "basePath": "/",
  "tags": [
    {
      "name": "Auth",
      "description": "Đăng ký, xác thực người dùng"
    },
    {
      "name": "Users",
      "description": "Get info, change profile"
    },
    {
      "name": "File",
      "description": "Upload, download file"
    }
  ],
  "paths": {
    "/auth/": {
      "parameters": [
        {
          "name": "Body",
          "required": true,
          "in": "body",
          "description": "Auth",
          "schema": {
            "default": {
              "username": "admin",
              "password": "123123123"
            }
          }
        }
      ],
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "routeWithId",
        "summary": "Login",
        "description": "Login",
        "produces": [
          "application/json"
        ],
        "responses": {
        }
      }
    },
    "/auth/register": {
      "parameters": [
        {
          "name": "Body",
          "required": true,
          "in": "body",
          "description": "Auth",
          "schema": {
            "default": {
              "username": "admin",
              "password": "123123123",
              "email": "email@yahoo.com"
            }
          }
        }
      ],
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "registerUser",
        "summary": "Register new user",
        "description": "Register new user",
        "produces": [
          "application/json"
        ],
        "responses": {
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/users/update": {
      "parameters": [
        {
          "name": "Body",
          "required": true,
          "in": "body",
          "description": "Auth",
          "schema": {
            "default": {
              "fullName": "Hoàng Ngọc Thạch",
              "email": "ngoisaocodon@gmail.com"
            }
          }
        }
      ],
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "updateUser",
        "summary": "Update profile",
        "description": "Update profile",
        "produces": [
          "application/json"
        ],
        "responses": {
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/users/info": {
      "parameters": [
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "operationId": "Get profile",
        "summary": "Get profile",
        "description": "Get profile",
        "produces": [
          "application/json"
        ],
        "responses": {
        }
      }
    },
    "/users/change-avatar": {
      "parameters": [
        {
          "type": "file",
          "in": "formData",
          "maxItems": 1,
          "name": "files"
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "operationId": "Get profile",
        "summary": "Get profile",
        "description": "Get profile",
        "produces": [
          "application/json"
        ],
        "responses": {
        }
      }
    }
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "Authentication token with Bearer token"
    }
  }
}